Add HTML boilerplate, title, and heading
To setup the webpage I decide to add the usual HTML boilerplate and formatted it so that the HTML is easier to read and edit. I gave the website a title so that it gives the website a readable title for others users who enter the website. The heading I add will also help users locate where all the recipes are.