Add HTML boilerplate title and header
Copied HTML boilerplate from the homepage (index.html). I changed the title so that it helps users locate their current place on the website. The new header will tell users that the website is about a macaroni and cheese recipe so that if a user made a mistake or wants to check if they are in the right place they can look at the heading.